Add options for patch library and library delegation before app class path
Adds config attribute called libraryDelegation to the global
classloading config that has options beforeApp and afterApp where
afterApp is the default behavior today.  When using beforeApp
option the library delegation will happen before the application
class path and before parent delegation.  The library loader will
also be configured as parentLast delegation.

Adds a config attribute on classloader config to called
patchLibraryRef which is similar to privateLibraryRef.  Both
options take the library class path content and place it
on the local class path of the application.  A patchLibrary
will add the content to the beginning of the local class path.
A privateLibrary adds the content to the end of the local class path