Actually Repeat Jakarta 3.0 tests for EE11
While the repeat logic caused the tests to run twice, due to the nature of how the configuration was managed, all that changed was EE11_FEATURES was added to the test and server names. the server configuration was left as is.

Use the RepeatTestFilter to determine if we are running the EE11 feature set and switch to the ee11 version of the configuration which includes the EE11 compatible feature set.