Actually Repeat Jakarta 3.0 tests for EE11
While the repeat logic caused the tests to run twice, due to the nature of how the configuration was managed, all that changed was EE11_FEATURES was added to the test and server names. the server configuration was left as is.

Use the RepeatTestFilter to determine if we are running the EE11 feature set and switch to the ee11 version of the configuration which includes the EE11 compatible feature set.

For the tests using the AuthenticationAppServlet, this test relied on Servlet-6.0 behaviour that we could still access the OutputStream despite having committed the response. In Servlet-6.1 the retrieval of the stream led to errors for EE11 tests. given what we were returning was not even sent back to the client, but simply logged. the code here now checks if the response has been committed and provides an appropriate output mechanism for logging purposes.